requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
fastapi==0.112.2
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typing_extensions==4.12.2
uvicorn==0.30.6
